+++ src/agentscope/tuner/_judge.py
@@ -1,0 +1,38 @@
+# -*- coding: utf-8 -*-
+"""A judge function type for tuning.
+
+Args:
+    task (Dict):
+        The task information for the corresponding workflow.
+    workflow_output (WorkflowOutput):
+        The output generated by the corresponding workflow.
+    auxiliary_models (Dict[str, ChatModelBase]):
+        A dictionary of additional chat models available for LLM-as-a-Judge
+        usage. The keys are model names, and the values are the corresponding
+        ChatModelBase instances.
+Returns:
+    JudgeOutput:
+        The reward value assigned by the judge function.
+"""
+from typing import Optional, Any, Callable, Dict, Awaitable
+from pydantic import BaseModel, Field
+from agentscope.model import ChatModelBase
+
+
+class JudgeOutput(BaseModel):
+    """The output of a judge function."""
+
+    reward: float = Field(
+        description="The reward value assigned by the judge function.",
+    )
+
+    metrics: Optional[Dict[str, float]] = Field(
+        description="Metrics from the judge function.",
+        default=None,
+    )
+
+
+JudgeType = Callable[
+    [Dict, Any, Dict[str, ChatModelBase]],
+    Awaitable[JudgeOutput],
+]
