+++ src/agentscope/tuner/_workflow.py
@@ -1,0 +1,52 @@
+# -*- coding: utf-8 -*-
+"""An agent workflow function type for tuning.
+
+Args:
+    task (Dict):
+        The task information for the workflow run.
+    model (ChatModelBase):
+        The primary chat model used in the workflow, this is the main model
+        being tuned.
+    auxiliary_models (Dict[str, ChatModelBase]):
+        A dictionary of additional chat models available for LLM-as-a-Judge
+        usage. The keys are model names, and the values are the corresponding
+        ChatModelBase instances. Note that these auxiliary models are not tuned
+        during the workflow.
+
+Returns:
+    WorkflowOutput:
+        The reward obtained from the workflow function.
+"""
+from typing import Optional, Any, Callable, Dict, Awaitable
+from pydantic import BaseModel, Field
+from agentscope.model import ChatModelBase
+
+
+class WorkflowOutput(BaseModel):
+    """The output of a workflow function."""
+
+    reward: Optional[float] = Field(
+        description=(
+            "The reward obtained from the workflow function. "
+            "Used for direct reward output."
+        ),
+        default=None,
+    )
+    response: Optional[Any] = Field(
+        description=(
+            "The response generated by the workflow function. "
+            "Used as judge input."
+        ),
+        default=None,
+    )
+
+    metrics: Optional[Dict[str, float]] = Field(
+        description="Metrics from the workflow function.",
+        default=None,
+    )
+
+
+WorkflowType = Callable[
+    [Dict, ChatModelBase, Dict[str, ChatModelBase]],
+    Awaitable[WorkflowOutput],
+]
